Learning ToBoolean
Referred the following links

ToBoolean : https://262.ecma-international.org/10.0/#sec-toboolean

Undefined→False
Null→False
For Number →if it is +0,-0 or NaN,it will return false,otherwise return true for all other values.
For String→If it is an empty string,it will return false,otherwise return true for all other values.

“Undefined,Null,NaN,+0,-0,false”-->These all are falsy values