Understanding Synchronous Nature of Javascript
JavaScript is Synchronous in Nature-->It is executes code line by line If we execute valid ECMA Script Code /Native.

JS is single threaded/single process.

We will never stop a synchronous part of code.

All the functions execute in synchronous nature on the call stack.
When the timer will end,all the processes inside setTimeOut function  will wait in the EVENT QUEUE untill synchronous/valid code is executed.JavaScript does not recognises the seTimeOut Function,It is recognised by the runtime.