Q8
Use this URL - https://example.com/api/itemlist to make a fake fetch call and list out all the items as an ordered list on the DOM. A fakeFetch has been provided. Use HTML, CSS & JS template in REPL or Vanilla template in CodeSandbox for this question.

Output on the DOM should be in the format, {itemName} -- INR {price} -- {quantity}:
  1. Bread -- INR 30 -- 10
  2. Water Bottle -- INR 50 -- 50
  3. Dairy Milk -- INR 20 -- 30